Rename docker package to generic `containerclient`
