+++ containerclient/dockerfile/testdata/moby.txt
